Propagated source values from params to destined module.
This could also have been done by only specifying the value in
hiera.yaml file. But, there is a possibility that one might forget
to configure this variable in hiera and the configuration will not
throw any errors. To avoid such misconfiguration, specifying that
value in params.pp so that it is always specified in hiera.yaml